Rework search UI.
Numerous improvements in response to feedback. The biggest
change that happened here is we are no longer narrowing
down the filter fields, so they are much simpler. This
wipes out a whole class of UI bugs. This will get us by
long enough to rework the UI with some nice Ajax.