Delete SQLite scripts.
We no longer need sqlite scripts because we can use the
orm-level dumpdata/loaddata tool to move data in a mostly
database agnostic way.

Added data.json, which has most of the non-user generated
data. This protects confidentiality, so we can store it in
the public repo.

We can now use "python manage.py loaddata utils/data.json"
to import a good chunk of recent data.