Add slugs to urls for better SEO.
Now supporting optional slugs in vendor URLs. By optional
I mean we should use them everywhere on our site, but that
they are not enforced.

For example:
/vendors/5/
/vendors/5-hello-world/
/vendors/5-hello-universe/

will all point to the same vendor, with id 5. I did some
research and this technique is acceptable. In fact, it is
used in some form by some large scale websites, like the
onion.