Fix zip code search.
Zip code searches were always returing no results. The reason for this
is that zip code searches, when run through the geocoding module, do
not come back with a neighborhood assigned, for whatever reason.
(This is subject to further review as it is believed that we are probably
not parsing google api JSON correctly.) Then, the search method on the
model manager was throwing away all results that did not have a
neighborhood. I am not sure why, but I think this is in response to
the ongoing problem that google would rather return a bad result
than no result. So, we were throwing away latlongs that had no
neighborhood because we thought they were just coming from bad queries.

Removing the block of code that throws away queries that don't return
neighborhoods.