Fix map centering, add color coded map icons.
Fix map centering so that when you come to the vendors page initially,
it centers over center city and ignores the outliers. There is a settings
constant that can be used to change this center for other regional sites.

Add color coding by veg level of map icons. Needs styling work, but should
be a nice standin.

Heavily refactor js. Create separate files for the two pages that use the
vendorMap object. Make underscore.js templates to handle rendering large
blobs of text better. Have js dependencies get pulled in as form.media
for the search_form on vendors.html.