Improve defaults for approval_status in admin interface
Improval status should not default to "approved", for anyone, ever.
That was a bad idea and we've learned from it. This was a bigger
problem in the data entry form processing view, but in this case
it's still a good idea to change it too.

Remove editable approval_status from summary page because it
just makes the page look busy and weird, and saving is not intuitive.

Reorder presentation on summary screen.