Change fonts and cleanup CSS
This should have been several separate commits but Ali is still
learning and making mistakes. So here's a massive commit to change
the fonts sitewide and to clean up the CSS by moving general
selectors to the top, removing comments, and moving page-specific
styling to internal stylesheets of each page.