Fix broken test '}}' => '};}' in inline js block
TravisCI build 77 failed because there is a test that checks all
rendered html pages for '{{' and '}}' as a way to make sure there are
no broken templatetags. This is a good test, but it raised a false
positive for inline javascript. Rather than make the test more robust,
it's easy enough to just add a semicolon into the js.

https://travis-ci.org/vegphilly/vegphilly.com/builds/9306103