Use a Backbone View to organize dom events
In the interest of incremental progress toward more organized js,
moving a lot of dom events that were unorganized in a document ready
block into a backbone view for more structure.

Obviously, a lot more can be done to make the JS more organized, like
using a model, using a require/module library, etc. This is a sep in
the right direction.