Fix bug in generic form processing view
views that were calling this view with commit=False so that they could
run some pre-save functions after saving were not ever being saved.

this introduces/exposes a bug in the veg user creation form, which is
that the email should be saved 'with' the rest of the form. Using a workaround
in the form of saving the object twice, quickly.