Expand dev fixture, stop using private_data
We were keeping a gpg'd private data file in the main repo for admin's
to decrypt and load. This was strange at best for an open source
project. Now, there are only 3 users in the database: vagrant/vagrant,
the superuser, dev_user_1, and dev_user_2. All user data has been
scrubbed so that all reviews are logged to one of these two
users. There's no need to keep this data private.