Initial commit of extended logging
fixes #112 on github.

Add logfiles for lots of stuff.

There are 3 logs for gunicorn, 3 for django (including db sql), and 2
for vegancity the app.

The app will now log all user searches, as an effort to start gathering
analytics about what people search for to improve the search
functionality of the site.

Realistically, some of these logs will have to be collapsed, removed, or
turned to log only warning+ severity messages. In the meantime, we've
got a lot of output to look at.

Modify tests to avoid logging, and to attach a user to the search view,
which is now required.