Read google analytics ID from settings file
Instead of using an overridable empty template for analytics, which is the
Wrong Way(tm). Now the JS will see if there's a tracking ID and only
initialize when there is one. This will work well with the
settings_local pattern.

this required adding a global context processor, but that will come in
handy anyway.